refactor: Navigation 컴포넌트 디렉토리 위치 변경 및 일부 네이밍 수정
